% Auto-generated: do not edit by hand
\name{mySilder}

\alias{mySilder}

\title{MySilder component}

\description{
ExampleComponent is an example component. It takes a property, `label`, and displays it. It renders an input with the property `value` which is editable by the user.
}

\usage{
mySilder(id=NULL, picUrl=NULL, step=NULL, value=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{picUrl}{Character. The value displayed in the input.}

\item{step}{Numeric. A label that will be printed when this component is rendered.}

\item{value}{Character | numeric | named list | unnamed list. The value displayed in the input.}
}
